update delete to spit out what was deleted

+++ xhrActions/deleteFile.js
@@ -36,7 +36,12 @@ module.exports = (request, response) => {
             logger.error('Error: ' + stringify(err));
             return;
         }
-        const successMsg = stringify(success);
+        const successMsg = `File ${filename} deleted!`;
+        if (filename.endsWith('.dng')) {
+            response.writeHead(200, {});
+            response.end('Done! ' + successMsg);
+            return;
+        }
         unlink(`${BASE_CONFIG_PATH}/${filename}.cfg`, (xerr, xsuccess) => {
             if (xerr) {
                 const message = 'Error: ' + stringify(xerr) + ' and success: ' + successMsg;
@@ -46,8 +51,9 @@ module.exports = (request, response) => {
                 return;
             }
 
+            const xsuccessMsg = `File ${filename}.cfg deleted!`;
             response.writeHead(200, {});
-            response.end('Done! ' + successMsg + ' and ' + stringify(xsuccess));
+            response.end('Done! ' + successMsg + ' ' + xsuccessMsg);
         });
     });
 };
